[ci-cd] Include AWS Amplify detection Jekyll makefile
Despite utilising the same Ruby OCI container/image the AWS Amplify
CRI differs to that of traditional, multi-purpose container engines such
as Docker and Podman.

AWS Amplify injects a set of environment variables at runtime:
https://docs.aws.amazon.com/amplify/latest/userguide/environment-variables.html.
The 'AWS_PLATFORM' environment variable was deemed sufficiently
self-documenting and therefore included as part of the makefile's
environment autodetection.

The existence of the 'AWS_PLATFORM' environment variable informs the
makefile it is within the desired operational context to invoke Jekyll
commands.